requirements.txt
azure-devops==6.0.0b4
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==37.0.0
fernet==1.0.1
Flask==2.0.2
Flask-Cors==3.0.10
future==0.18.2
httplib2==0.20.2
idna==3.3
IMAPClient==2.2.0
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
MouseInfo==0.1.3
msrest==0.6.21
oauthlib==3.2.0
pefile==2021.9.3
py2exe==0.11.0.1
pyaes==1.6.1
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-decouple==3.6
pytweening==1.0.4
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
socks==0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.3
XlsxWriter==3.0.2
